+++ LocationCompass/Properties/AssemblyInfo.cs
