+++ LocationCompass/constants.json
