+++ LocationCompass/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Pathoschild.Stardew.ModBuildConfig" version="2.2.0-beta-20180819" targetFramework="net45" />
+</packages>
